rename wsgi folders
