+++ label_studio_ml/examples/watsonx_llm/_wsgi.py
@@ -1,7 +1,7 @@
 from werkzeug.middleware.dispatcher import DispatcherMiddleware
 
-from wsgi_data._wsgi import application as data
-from wsgi_model import _wsgi as model
+from data_wsgi._wsgi import application as data
+from model_wsgi import _wsgi as model
 
 
 app = DispatcherMiddleware(model.application, {
