+++ label_studio_ml/examples/watsonx_llm/model_wsgi/model.py
